Update and rename Task2.cpp to Task02.cpp
Do not make functions for printing strings, just `std::cout << str << std::endl`
